+++ model/task_queue_service.go
@@ -20,6 +20,30 @@ type CachedDispatcher interface {
 	Type() string
 }
 
+type dispatchSkipsKey struct{}
+
+// NewDispatchSkipsContext returns a context that counts the number of tasks
+// skipped in a next_task attempt.
+func NewDispatchSkipsContext(ctx context.Context) context.Context {
+	return context.WithValue(ctx, dispatchSkipsKey{}, new(int))
+}
+
+// RecordDispatchSkip increments the context's dispatch skip counter.
+func RecordDispatchSkip(ctx context.Context) {
+	if count, ok := ctx.Value(dispatchSkipsKey{}).(*int); ok {
+		*count++
+	}
+}
+
+// DispatchSkipsFromContext returns the number of dispatch skips recorded in
+// the context.
+func DispatchSkipsFromContext(ctx context.Context) int {
+	if count, ok := ctx.Value(dispatchSkipsKey{}).(*int); ok {
+		return *count
+	}
+	return 0
+}
+
 type taskDispatchService struct {
 	cachedDispatchers map[string]CachedDispatcher
 	mu                sync.RWMutex
